delete comments, refactoring

+++ resources/views/ad/index.blade.php
@@ -21,31 +21,19 @@
     <br/>
     <div class="container">
         <div class="form-group">
-          {{--  <a href="{{ url('ads') }}" class="btn btn-success">Logout</a>--}}
             <a href="{{ route('logout')}}" class="btn btn-success">Logout</a>
-
-
-            {{--<a href="{{ route('register')}}" class="btn btn-success">Logout</a>--}}
             <hr>
         </div>
     </div>
 
-    @else
+@else
     <form action="/register" class="form-signin" method="post">
         {{csrf_field()}}
         <h1 class="h3 mb-3 font-weight-normal">Please sign in</h1>
         <label for="inputUserName" class="sr-only">User name</label>
-        {{--<input type="email" id="inputEmail" class="form-control" placeholder="User name" required autofocus>--}}
         <input type="username" id="inputUserName" name="username" class="form-control" align="center"  placeholder="User name" required autofocus>
         <label for="inputPassword" class="sr-only">Password</label>
         <input type="password" id="inputPassword" name="password" class="form-control" placeholder="Password" required>
-        {{--<label for="confirmPassword" class="sr-only">Confirm Password</label>
-        <input type="password" id="confirmPassword" name="password_confirmation" class="form-control" placeholder="Confirm Password" required>--}}
-{{--        <div class="checkbox mb-3">
-            <label>
-                <input type="checkbox" name="remember" value="1"> Remember me
-            </label>
-        </div>--}}
         <button class="btn btn-lg btn-primary btn-block" type="submit">Sign in</button>
         <p class="mt-5 mb-3 text-muted">&copy; 2018</p>
     </form>
@@ -53,21 +41,8 @@
 </html>
 @endif
 
-
 @extends('layout.template')
 @section('content')
-
-
-
-    {{--<div class="container">
-      <div class="form-group">
-              <h1>List Ads:</h1>
-              <a href="{{url('/ads/create')}}" class="btn btn-success">Create Ad</a>
-              <hr>
-      </div>
-      </div>--}}
-
-
 
     @if(Auth::check())
         <div class="container">
@@ -79,37 +54,8 @@
         </div>
     @endif
 
-    {{--правило не срабатывает--}}
-
-    {{--/*        if($request->user()->cannot('add', $article)){
-    return redirect()->back()->with(['message'=>'У Вас нет прав']);
-    }*/
-
-    /*  ???      if(Gate::denies('add', $ad)){
-    return redirect()->back()->with(['message'=>'У Вас нет прав.']);
-    ???return redirect()->back();
-    }*/--}}
+    {{--the @can rule does not work--}}
     @can('add', $user)
-    <div class="container">
-        <div class="form-group">
-            <h1>List Ads:</h1>
-            <a href="{{url('/ads/create')}}" class="btn btn-success">Create Ad!!!</a>
-            <hr>
-        </div>
-    </div>
-    @endcan
-
-    @can('update', $ads)
-    <div class="container">
-        <div class="form-group">
-            <h1>List Ads:</h1>
-            <a href="{{url('/ads/create')}}" class="btn btn-success">Create Ad!!!</a>
-            <hr>
-        </div>
-    </div>
-    @endcan
-
-    {{--@if(Gate::allow('add','ads'))
         <div class="container">
             <div class="form-group">
                 <h1>List Ads:</h1>
@@ -117,8 +63,28 @@
                 <hr>
             </div>
         </div>
-    @endif--}}
-    {{--правило не срабатывает--}}
+    @endcan
+
+    @can('add')
+        <div class="container">
+            <div class="form-group">
+                <h1>List Ads:</h1>
+                <a href="{{url('/ads/create')}}" class="btn btn-success">Create Ad!!!</a>
+                <hr>
+            </div>
+        </div>
+    @endcan
+
+    @can('update', $ads)
+        <div class="container">
+            <div class="form-group">
+                <h1>List Ads:</h1>
+                <a href="{{url('/ads/create')}}" class="btn btn-success">Create Ad!!!</a>
+                <hr>
+            </div>
+        </div>
+    @endcan
+    {{--the @can rule does not work--}}
 
 
 
@@ -155,18 +121,15 @@
                 @if($item->user_id == $user->id)
                     <td><a href="{{route('ads.edit',$item->id)}}" class="btn btn-warning">Edit</a></td>
                     <td>
-                        {{--  {!! Form::open(['method' => 'DELETE', 'route'=>['ads.destroy', $item->id]]) !!}
-                          {{ csrf_field() }}
-                          {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                          {!! Form::close() !!}--}}
 
-                        {{--                    @can('delete', $ads)
-                                                {!! Form::open(['method' => 'DELETE', 'route'=>['ads.destroy', $item->id]]) !!}
-                                                {{ csrf_field() }}
-                                                {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                                                {!! Form::close() !!}
-                                            @endcan--}}
-
+                            {{--the @can rule does not work--}}
+                            @can('delete', $ads)
+                                {!! Form::open(['method' => 'DELETE', 'route'=>['ads.destroy', $item->id]]) !!}
+                                {{ csrf_field() }}
+                                {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                                {!! Form::close() !!}
+                            @endcan
+                            {{--the @can rule does not work--}}
 
                         {!! Form::open(['method' => 'DELETE', 'route'=>['ads.destroy', $item->id]]) !!}
                         {{ csrf_field() }}
